[common] np.concatenate

+++ TrainAndTest/train_egs_distributed.py
@@ -799,7 +799,7 @@ def main():
 
                 # valid_test(train_extract_loader, model, epoch, xvector_dir)
                 test(extract_loader, model, epoch, writer, xvector_dir)
-                if epoch != (end - 1):
+                if epoch != (end - 1) and torch.distributed.get_rank() == 0:
                     try:
                         shutil.rmtree("%s/train/epoch_%s" % (xvector_dir, epoch))
                         shutil.rmtree("%s/test/epoch_%s" % (xvector_dir, epoch))
